requirements/dev-requirements.txt
-r requirements.txt

flake8==3.9.2
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
model-bakery==1.3.2
djangorestframework-simplejwt
